feat: prefer h264 codec for youtube download (#535)
I was trying around with downloading karaoke-videos from youtube. 
Often these are encoded in h265 or vp9 or av1. 
The pi really struggles with formats like that, because they need
additional cpu resources.

After setting a parameter for yt-dlp the the codes h264 is preferred.
This helped alot when playing videos on a high-resolution tv.