fix: Update yt-dlp to 2025.11.12, add js runtime support (#541) (#553)
yt-dlp will soon require deno (or other JavaScript runtime). This PR
adds the relevant package and setup script to Dockerfile so that the
docker image contains those dependencies.

This is very unfortunate, because it will require more external
dependencies that need to be installed manually. What's more, yt-dlp
prefers deno as the default installation. It will not recognize the much
more common node runtime unless we tack on `--ytdl-args "--js-runtimes
node"`

FWIW Currently I am not detecting any issues with downloading videos
without the runtime, despite the warnings in the logs.

---------

Co-authored-by: cdytoby <cdytoby@live.cn>