feat: personalized score phrases
I've added the ability for users to create personalized phrases for the score screen.

Why? Besides just adding more phrases, when bringing your system to a party with friends and family, you can create custom phrases with names and inside jokes that will make the moment more fun.

Also fixed the “checked” state on the checkboxes on the info page, which wasn’t rendering correctly.